abstract base shell class
